Move auth replace mint/burn to a separate file
